Fix terminology used to refer to REST API version
There is no longer a concept of a "major" version; as of protocol
version 2, the protocol version is a single integer (see spec point
CSV2a).

Unfortunately, the linked ably.com REST API versioning documentation
still refers to API versions 1.1 and 1.2, but that’s a separate issue —
the REST API documentation needs to be versioned and updated for newer
protocol versions, which Mark says is planned.

Relevant conversation in [1], where Mark says:

> For the docstring I think we can drop "major" and just refer to it as
> "version" since this will be the case going forwards. We can leave the
> link in for now too even if it'll be a little wonky until we release the
> versioned REST API.

[1] https://github.com/ably/ably-js/pull/1670#discussion_r1532211929